.util/requirements.txt
gitpython
python-dotenv
httpx
click
google-api-python-client
google-cloud-secret-manager==0.1.1
